# aifo/slidescope/internal/server/public/openapi.yaml
basePath: /api/v1
definitions:
  domain.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  domain.PhysicalDimensions:
    properties:
      heightMm:
        type: number
      heightPx:
        type: integer
      widthMm:
        type: number
      widthPx:
        type: integer
    type: object
  domain.Slide:
    properties:
      slideHeight:
        type: integer
      slideId:
        type: string
      slideMpp:
        type: number
      slideName:
        type: string
      slideUri:
        type: string
      slideWidth:
        type: integer
    type: object
  domain.SlideMetadata:
    properties:
      format:
        type: string
      magnification:
        type: string
      maxLevel:
        type: integer
      minLevel:
        type: integer
      physicalDimensions:
        $ref: '#/definitions/domain.PhysicalDimensions'
      slideHeight:
        type: integer
      slideId:
        type: string
      slideMpp:
        type: number
      slideName:
        type: string
      slideWidth:
        type: integer
      tileSize:
        type: integer
      vendor:
        type: string
    type: object
  domain.SlidesResponse:
    properties:
      slides:
        items:
          $ref: '#/definitions/domain.Slide'
        type: array
    type: object
  domain.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
  domain.User:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  handlers.LoginInput:
    properties:
      password:
        example: password123
        type: string
      username:
        example: testuser
        type: string
    required:
      - password
      - username
    type: object
  handlers.SlideInput:
    properties:
      slideId:
        example: slide123
        type: string
      slideName:
        example: Sample Slide
        type: string
      slideUri:
        example: file:///path/to/slide.svs
        type: string
    required:
      - slideUri
    type: object
  handlers.UserInput:
    properties:
      password:
        example: password123
        type: string
      username:
        example: testuser
        type: string
    required:
      - password
      - username
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A digital pathology slide management and viewing API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SlideScope API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
        - application/json
      description: Authenticate user credentials and return JWT tokens in cookies
      parameters:
        - description: User credentials
          in: body
          name: credentials
          required: true
          schema:
            $ref: '#/definitions/handlers.LoginInput'
      produces:
        - application/json
      responses:
        "200":
          description: Authentication successful
          schema:
            $ref: '#/definitions/domain.TokenResponse'
        "400":
          description: Bad request - invalid input
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized - invalid credentials
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: User login
      tags:
        - authentication
  /api/v1/auth/logout:
    post:
      description: Clear authentication cookies and log out the user
      produces:
        - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            additionalProperties:
              type: string
            type: object
      security:
        - BearerAuth: []
      summary: User logout
      tags:
        - authentication
  /api/v1/auth/me:
    get:
      description: Retrieve information about the currently authenticated user
      produces:
        - application/json
      responses:
        "200":
          description: Current user information
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
        - BearerAuth: []
      summary: Get current user information
      tags:
        - authentication
  /api/v1/auth/refresh:
    post:
      description: Use refresh token to generate new access and refresh tokens
      produces:
        - application/json
      responses:
        "200":
          description: Token refresh successful
          schema:
            $ref: '#/definitions/domain.TokenResponse'
        "401":
          description: Unauthorized - invalid refresh token
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Refresh authentication tokens
      tags:
        - authentication
...
  /api/v1/slides:
    get:
      description: Retrieve a list of all available slides
      produces:
        - application/json
      responses:
        "200":
          description: List of slides
          schema:
            $ref: '#/definitions/domain.SlidesResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
        - BearerAuth: []
      summary: Get all slides
      tags:
        - slides
    post:
      consumes:
        - application/json
      description: Add a new slide to the system
      parameters:
        - description: Slide information
          in: body
          name: slide
          required: true
          schema:
            $ref: '#/definitions/handlers.SlideInput'
      produces:
        - application/json
      responses:
        "201":
          description: Created slide
          schema:
            $ref: '#/definitions/domain.Slide'
        "400":
          description: Bad request - invalid input
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
        - BearerAuth: []
      summary: Add a new slide
      tags:
        - slides
  /api/v1/slides/{slide_id}:
    get:
      description: Retrieve a specific slide by its unique identifier
      parameters:
        - description: Slide ID
          example: '"slide123"'
          in: path
          name: slide_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Slide information
          schema:
            $ref: '#/definitions/domain.Slide'
        "400":
          description: Bad request - slide_id required
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Slide not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
        - BearerAuth: []
      summary: Get slide by ID
      tags:
        - slides
  /api/v1/slides/{slide_id}/metadata:
    get:
      description: Retrieve detailed metadata and properties for a specific slide
      parameters:
        - description: Slide ID
          example: '"slide123"'
          in: path
          name: slide_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Slide metadata
          schema:
            $ref: '#/definitions/domain.SlideMetadata'
        "400":
          description: Bad request - slide_id required
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Slide not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
        - BearerAuth: []
      summary: Get slide metadata
      tags:
        - slides
...
  /api/v1/users:
    post:
      consumes:
        - application/json
      description: Create a new user account with username and password
      parameters:
        - description: User credentials
          in: body
          name: user
          required: true
          schema:
            $ref: '#/definitions/handlers.UserInput'
      produces:
        - application/json
      responses:
        "201":
          description: Created user (password field will be empty)
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad request - invalid input
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
        - BearerAuth: []
      summary: Create a new user
      tags:
        - users
  /api/v1/users/{username}:
    get:
      description: Retrieve user information by username
      parameters:
        - description: Username
          example: '"testuser"'
          in: path
          name: username
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: User information
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad request - username required
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
        - BearerAuth: []
      summary: Get user by username
      tags:
        - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"